academic_research_assistant/requirements.txt
google-adk
fastapi
uvicorn[standard]
python-dotenv
pymupdf
semanticscholar
requests
pydantic
